reviewService 버error fix

+++ src/main/java/com/example/ajouparking/service/ReviewService.java
@@ -44,7 +44,7 @@ public class ReviewService {
     @Transactional
     public ResponseEntity<?> saveReview(long userId, Long parkingLotId, ReviewRequestDto reviewRequestDto) {
         User user = userRepository.findById(userId).orElseThrow(()->new IllegalArgumentException("없는유저"));
-        int reviewAuthorId = reviewRequestDto.getUser().getId();
+        long reviewAuthorId = reviewRequestDto.getUser().getId();
         if(reviewAuthorId == userId){
             throw new CustomApiException("자신의 리뷰에 좋아요 할 수 없습니다");
         }
